Cleaned up headers
