+++ Board/Bitboards/bitboard.h
@@ -10,5 +10,6 @@ void resetBit(u64 *bb, int index);
 
 int bitScanForward(u64 bb);
 int popls1b(u64 *bb);
-u64 signedShift(u64 bb, int s);
 int bitcount(u64 bb);
+
+u64 signedShift(u64 bb, int s);
