+++ Search/eval.h
@@ -1,4 +1,5 @@
-#include "../types.h"
+#pragma once
+
 #include "../Board/board.h"
 
 double evaluate(Board *board);
