+++ Search/movegen.c
@@ -1,5 +1,8 @@
 #include "movegen.h"
 
+#include "Magic/magic.h"
+#include "../Board/Bitboards/bitboard.h"
+
 u64 knightMoves[64];
 u64 kingMoves[64];
