+++ ui/print.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include "../Board/board.h"
 #include "../types.h"
 
@@ -9,13 +10,13 @@ struct printSettings{
 };
 
 typedef struct printSettings printSettings;
+
 //set light and dark color will result in \x1b[colorIDm
 void setLightColor(printSettings *settings, const char *colorID);
 void setDarkColor(printSettings *settings, const char *colorID);
-
 void setUnicodePieces(printSettings *settings);
 void setASCIIPieces(printSettings *settings);
-void printBoard(printSettings settings, Board board, u64 highlighted);
 
+void printBoard(printSettings settings, Board board, u64 highlighted);
 void printMoveOnBoard(printSettings settings, Board board, Move m);
 void printBitboardOnBoard(printSettings settings,Board board,u64 bb);
