--- ui/tostr.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <stdio.h>
-#include <string.h>
-#include "../types.h"
-#include "../Board/board.h"
-
-void moveToString(Move move, char *target);
